Move map layers to their own folder
The DrawMap components folder is starting to get a lot of files. Since
there will be more layer components in the future, it would be good to
split them out now.
