+++ src/app/src/pages/Welcome.js
@@ -1,6 +1,6 @@
 import Map from '../components/Map';
 import WelcomeModal from '../components/WelcomeModal';
-import { DefaultBasemap } from '../components/Basemaps';
+import { DefaultBasemap } from '../components/Layers/Basemaps';
 
 export default function Welcome() {
     return (
